feat(*): Update quoted text style
This commit updated the quoted text style app wide. Before this commit
all the quoted text in notice and post are wrapped in a card with more
elevation. From this commit on, these quoted text are surrounded by icon
which looks like double quote symbol.
Added `QuotedText` widget with two ways to construct:
* Construct from normal `String` type text, will wrap in a `RichText`.
* Construct from an `InlineSpan`, will be wrapped in a row with two
  `RichText` to ensure the visible style.