feat(cache): Support reloading cached images
This commit added support for reloading cached images. User can manually
trigger image refreshing in image action bottom sheet.

The new feature is achieved by using a global cache manager repository
which expose a stream of events (called response in code) describe
events happened on images. Different images are distinguished by image
url. And in `CachedImage`, the widget is listening to the stream so that
rebuild itself according to image events.

`CachedImage` is supported though `CachedImageProvider` is not yet.

This commit also refatored naming of `CacheRepository` used in settings
feature.