fix(*): Fix wrong target thread page from notice
This commit fixes an issue where user was headed to wrong thread page
number when redirect to thread page from notice detail page, introduced
in e7b3819

Ths issue occurred when thread original order and app defined override
thread order are not the same, the opened page intended to be in the
original order but once override user got another page which did not
contains the target post.

So make `reverseOrder` flag additional and set it to null when pushing
routes from notification detail page.