fix(storage): fix notice table primary key
In previous version, column nid in notice table is mis-ignored and
timestamp is a part of primary key instead. But when notices generated
by batch rating are having the same timestamp.

This commit updated the primary key in notice table to `{uid, nid}`
which distinguish notice with its id, not timestamp.