feat(editor): roll editor and enable resizing image appearance
This commit imeplemented the resize of appearance when user changed
image size in bbcode. Now image size saved in bbcode is applied on
the related image not only when generating bbcode when posting data
to server, but also updated the appearance of image in the editor,
providing a consist look and feel in editor and web side.

This commit also added a image resizing logic to act like web side:

* The maximum width of image on the web side is 550 currently.
* If an image has a larger width (in bbcode) than the maximum limit,
  image is scaled down to meet the max width while keeping the width
  height ratio.
* If an iamge has a smaller width (in bbcode) than the maximum limit,
  image is rendered with size in bbcode, no matter the height value.