fix(storage): fix potential migrate issue from 4 to 5
This commit fixed an potential migration issue on the notice table
in the migration from 4 to 5. The original migrate step uses some
API to deserialize table query row data to `NoticeEntity` but the
entity class is always the latest one which will cause an exception
and abort the migration. Current solution uses custom query to delete
all duplicate data which will block altering table and then use alter
api to change primary key.