feat(notification): scheduled pull actions, part 2
This commit made all cards in notification page work in read/unread mode,
all unread notice will show a badge on the leading in card, user can
dismiss them by tapping on notice card or url in it.

Till now all kinds of notice are supported in this behavior.

There are some limitations:

* Only personal messages' read/unread are synced through devices as
  server provides it. Means all other types of messages will keep in
  unread state even user already read them in another app instance.
* Badge on notice card may be mis-dismissed by tapping on urls unrelated
  to the notice. In a notice body, there may be more than one url that
  user can interact with and only part of them are the ones linke to
  current notice while others are unrelated, e.g. user profile url. So
  when user tapping on those unrelated urls, the badge also dimiss. This
  is impossible to fix without parsing the intent of notice: whether is
  linked to user profile or not, in the example above.