chore(editor): roll editor to support editing spoiler blocks
This commited rolled bbcode editor to latest revision to achive support
for editing spoiler block. A spoiler block in the editor now represents
like a regualr spoiler card in the html muncher. To edit a spoiler,
editor opens a new page and pop new content back.

Note that the conversion on spoiler block from/to quill delta and bbcode
is unimplemented yet.