perf(authentication): do not fetch user profile from settings page.
This commit removed the process of fetching full user info from user
password settings url.

Since the time multi-user target landed, user email is no longer part of
user credential, was removed from both authentication process and
storage. The full user info fetch process when login with document is
only intended to fetch the user email field: only that page shows it. So
it's time to remove it from the login process to make it faster.