fix(cached_image): remove loading splash
This commit removed the loading splash, now only the first load of image
has the splash and no one on later ones.

Achived by some magic not checked yet in the `Image` widget. Seems when
using `ImageProvider`, Image still keeps some cache in memory and avoid
loading it again if not necessary which is MUCH BETTER than the way
using `Image.raw`.

Now `CachedImage` is using `CachedImageProvider`, make it consist and
easier to maintain.
But maybe this commit breaks the feature of loading avif format images
as not following `ImageCacheBloc`. This should be checked and fixed if
necessary.

Fixes #18 #52