feat(*): reloadable images
This commit restored the functionality to

* Reload an image if requested, aka performing a force reload.
* Reload error ended user avatar if avatar refreshed and succeeded
  without the need of manually refreshing UI.

All image cached features are mainly achived in `ImageCachedProvider`.

Also removed the unused `ImageCacheBloc`, now image cached stream
comminication is done with the one exposed by `ImageCachedProvider`
which handling the process of loading images.

Note that to specify the usage of images, new type `ImageCacheRequest`
and `ImageCacheResponseType` are added to the stream. Naming is
confusing but fine.