fix(thread): find thread title by id
All thread with or without thread type has a `a` node with id
`thread_subject`, the node is visible in 爱丽丝 style and invisible in
other styles.

This commit use the id to select thread title to fix title not found
exception. Seems server changed the style recently which breaks the
old implementation.