refactor(thread): define thread bloc interface
This commit is part of migration of thread retrieving logic, defines
the common methods should be implmented by both existing v1 and planned
v2 implementations. Unfortunately the interfaces are derived by
`extends` not `implements` because the base class extends `Bloc` where
the we have to implement all methods in `Bloc` class if derive with
`implements`.

And note that all events are kept the same and state need to more
consideration, at the worst case, we can not keep a same basic bloc.