fix(router): remove top level single shell route
The singleton `ShellRoute` wrapped the a new navigator which makes
the a different one below the orignal root navigator. This breaks
all dialog pushing behavior result in the issues in previous commits.

Move the global singleton to the builder method so that we do can remove
the wrapper `ShellRoute`. But another caution is that the builder is
literally a `TransitionBuilder` not intended to wrap normal widget around
it and the function parameter is nullable. Keep checking the behavior at
least till next release.