feat: readonly editor not works
This commit added a way to build readonly bbcode editor. But contents
not rendered correctly.

Now the convertion from html to quill_delta haven't been implemented,
it's not possible to render into the right format till we find a way
to add flexible and ignored widgets in quill_delta, otherwise the
process steps will couple with the implementation in dependency.