fix: load an empty placeholder image instead of throwing
This commit is another workaround for the possible and expected
rendering failure in image provider. There is nothing to render if
the `CachedImageProvider` did not find the target image cache but
in Flutter's image provider API the rendering process still goes
further.

In previous versions, the error reporting was done by throwing new
exceptions, which is expensive and annoy the debugging process.
In this commit, the process is turned to be an as small as possible
image to avoid throwing exceptions. Things got better in some way
although it is not the best solution.