feat(reply_bar): parse content when editor reopen
This commit enabled the bbcode parsing step when bbcode editor in reply
bar is reopened, so that user still get the rendered result when the
reply bar is collapsed and expanded again.

But now this process is till for testing, because:

* BBCode parsing and rendering steps are done in `initState`
  synchronously, it may cause UI jank if the content is not short,
  beside when known performance issue in upstream quill_editor.
* The render process willl discard, more exactly keep plain text style,
  on styles that not supported by the bbcode parser, but at least this
  behavior is the same with what we have when editing posts.

Possible next step is keeping the constructed quill delta outside of the
lifetime of rich editor and pass that delta data in rich editor when it
reopened, instead of raw text, so we could make better performance and
avoid the second issue mentioned above. In addition a switch in thread
page in preferred.