fix: do not declare gzip encoding when findpost
In the mensioned issue in comment, a 301 status code from server causes
http handshake failure if gzip is declared in accept-encoding header.

Though server still responds gzip format data, it shall not exists in
header.

Not clear what is the root cause, the only workaround is removing gzip
encoding whereever a 301 status code is intended to receive.

Currently 'goto=findpost' requests are excluded in this commit.