feat: enable Impeller and add workaround for broken image
This commit brings Impeller back to Android again.

Specific image is broken with Impeller backend, no similar issues found
on issue tracker. To prepare for the upcoming stable release in August,
enable impeller on Android is required. For the broken image issue, till
now we only found the redirect back icon in static server resource is
affected, so this commit adds a workaround for it: instead of render the
original image, replace it with icon and icon button. This change is
what we DO NOT intend but have to, another possible workaround is on the
cache layer, but it is less semantic for users.